chore: setting up for mobile package nav
