+++ src/components/CatalogSearch/CatalogSearch.tsx
@@ -1,98 +1,18 @@
-import { Button, Grid, Input } from "@chakra-ui/react";
-import type {
-  ChangeEventHandler,
-  FormEventHandler,
-  FunctionComponent,
-} from "react";
-import {
-  Language,
-  TEMP_SUPPORTED_LANGUAGES,
-  LANGUAGE_NAME_MAP,
-} from "../../constants/languages";
-import { createTestIds } from "../../util/createTestIds";
+import { Grid } from "@chakra-ui/react";
+import type { FormEventHandler, FunctionComponent } from "react";
 import { Form } from "../Form";
-import { Dropdown, DropdownProps } from "./Dropdown";
+import {
+  CatalogSearchInputs,
+  CatalogSearchInputsProps,
+} from "./CatalogSearchInputs";
+import { testIds } from "./constants";
 
-type LanguageItems = Partial<Record<Language, string>>;
-
-const languageOptions = Object.fromEntries(
-  Object.entries(LANGUAGE_NAME_MAP).filter(([key]) =>
-    TEMP_SUPPORTED_LANGUAGES.includes(key as Language)
-  )
-) as LanguageItems;
-
-const LanguageDropdown: FunctionComponent<DropdownProps<LanguageItems>> =
-  Dropdown;
-
-export const testIds = createTestIds("catalog-search", [
-  "form",
-  "input",
-  "languageDropdown",
-  "languageDropdownMenu",
-  "languageDropdownValue",
-  "languageItem",
-  "submit",
-] as const);
-
-export interface CatalogSearchProps {
-  /**
-   * Controls the query state value
-   */
-  query: string;
-  /**
-   * Controls the query state change event
-   */
-  onQueryChange: ChangeEventHandler<HTMLInputElement>;
-  /**
-   * Controls the language state value
-   */
-  language: Language | null;
-  /**
-   * Controls the language state change event
-   */
-  onLanguageChange: (language: Language | null) => void;
+export interface CatalogSearchProps extends CatalogSearchInputsProps {
   /**
    * Called when the catalog search form is submitted (via enter keypress or submit click)
    */
   onSubmit: FormEventHandler<HTMLFormElement>;
 }
-
-export const CatalogSearchInputs: FunctionComponent<
-  Omit<CatalogSearchProps, "onSubmit">
-> = ({ query, onQueryChange, language, onLanguageChange }) => (
-  <>
-    <Input
-      bg="white"
-      borderColor="blue.100"
-      boxShadow="base"
-      data-testid={testIds.input}
-      name="query"
-      onChange={onQueryChange}
-      placeholder="Search providers or modules..."
-      value={query}
-    />
-    <LanguageDropdown
-      items={languageOptions}
-      onSelect={onLanguageChange}
-      placeholder="Language..."
-      selected={language}
-      testIds={{
-        item: testIds.languageItem,
-        menu: testIds.languageDropdownMenu,
-        trigger: testIds.languageDropdown,
-        value: testIds.languageDropdownValue,
-      }}
-    />
-    <Button
-      boxShadow="base"
-      colorScheme="blue"
-      data-testid={testIds.submit}
-      type="submit"
-    >
-      Search
-    </Button>
-  </>
-);
 
 export const CatalogSearch: FunctionComponent<CatalogSearchProps> = ({
   onSubmit,
