+++ src/components/CatalogSearch/constants.ts
@@ -1,0 +1,11 @@
+import { createTestIds } from "../../util/createTestIds";
+
+export const testIds = createTestIds("catalog-search", [
+  "form",
+  "input",
+  "languageDropdown",
+  "languageDropdownMenu",
+  "languageDropdownValue",
+  "languageItem",
+  "submit",
+] as const);
