+++ src/components/CatalogSearch/index.ts
@@ -1,1 +1,3 @@
 export * from "./CatalogSearch";
+export * from "./constants";
+export * from "./CatalogSearchInputs";
