+++ src/views/Package/components/LanguageSelection/LanguageSelection.tsx
@@ -1,4 +1,3 @@
-import { Flex } from "@chakra-ui/react";
 import type { Assembly } from "jsii-reflect";
 import type { FunctionComponent } from "react";
 import { LanguageBar } from "../../../../components/LanguageBar";
@@ -8,7 +7,6 @@ import {
   TEMP_SUPPORTED_LANGUAGES,
 } from "../../../../constants/languages";
 import { useLanguage } from "../../../../hooks/useLanguage";
-import { UseConstruct } from "../UseConstruct";
 
 export interface LanguageSelectionProps {
   assembly: Assembly;
@@ -29,18 +27,13 @@ export const LanguageSelection: FunctionComponent<LanguageSelectionProps> = ({
   ] as Language[];
 
   return (
-    <Flex align="center" justify="space-between">
-      <Flex direction="column">
-        <LanguageBar
-          selectedLanguage={targets.includes(language) ? language : targets[0]}
-          setSelectedLanguage={setLanguage}
-          showDisabled
-          targetLanguages={targets.filter((target) =>
-            TEMP_SUPPORTED_LANGUAGES.includes(target)
-          )}
-        />
-      </Flex>
-      <UseConstruct packageName={assembly.name} version={assembly.version} />
-    </Flex>
+    <LanguageBar
+      selectedLanguage={targets.includes(language) ? language : targets[0]}
+      setSelectedLanguage={setLanguage}
+      showDisabled
+      targetLanguages={targets.filter((target) =>
+        TEMP_SUPPORTED_LANGUAGES.includes(target)
+      )}
+    />
   );
 };
