+++ src/views/Package/components/PackageDetails/PackageDetails.tsx
@@ -1,9 +1,10 @@
-import { Center, Divider, Grid, Spinner } from "@chakra-ui/react";
+import { Center, Divider, Flex, Grid, Spinner } from "@chakra-ui/react";
 import type { Assembly } from "jsii-reflect";
 import { FunctionComponent } from "react";
 import type { Metadata } from "../../../../api/package/metadata";
 import { Card } from "../../../../components/Card";
 import type { UseRequestResponse } from "../../../../hooks/useRequest";
+import { LanguageSelection } from "../LanguageSelection";
 import { OperatorArea } from "../OperatorArea";
 import { PackageHeader } from "../PackageHeader";
 
@@ -33,23 +34,27 @@ export const PackageDetails: FunctionComponent<PackageDetailsProps> = ({
   }
 
   return (
-    <Grid
-      as={Card}
-      gap={4}
-      templateColumns={["1fr", null, "3fr auto 2fr"]}
-      templateRows="auto"
-    >
-      <PackageHeader
-        description={assembly.data.spec.description}
-        tags={assembly.data.spec.keywords ?? []}
-        title={assembly.data.spec.name}
-      />
-      <Divider display={["none", null, "initial"]} orientation="vertical" />
-      <Divider
-        display={["initial", "initial", "none"]}
-        orientation="horizontal"
-      />
-      <OperatorArea assembly={assembly.data} metadata={metadata.data} />
-    </Grid>
+    <Flex as={Card} direction="column">
+      <Grid
+        gap={4}
+        templateColumns={["1fr", null, "3fr auto 2fr"]}
+        templateRows="auto"
+      >
+        <PackageHeader
+          description={assembly.data.spec.description}
+          tags={assembly.data.spec.keywords ?? []}
+          title={assembly.data.spec.name}
+        />
+        <Divider display={["none", null, "initial"]} orientation="vertical" />
+        <Divider
+          display={["initial", "initial", "none"]}
+          orientation="horizontal"
+        />
+        <OperatorArea assembly={assembly.data} metadata={metadata.data} />
+      </Grid>
+      <Flex justify={["center", null, "start"]} px={2} py={4}>
+        <LanguageSelection assembly={assembly.data} />
+      </Flex>
+    </Flex>
   );
 };
